Koshucode.Baala.Minimal.OpKit

Description

Kit for implementing relational operators

Monoid

mappend :: Monoid a => a -> a -> a

An associative operation

mconcat :: Monoid a => [a] -> a

Fold a list using the monoid. For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

Abort

data Abort

Abort symbols

Constructors

AbortLookup [SourceLine] String

AbortMalformedOperand [SourceLine] String

AbortMalformedTerms [SourceLine] String

AbortMissingTermName [SourceLine] String

AbortUnknownClause [SourceLine]

AbortUnknownRelmap [SourceLine] String

AbortUsage [SourceLine] [String]

type AbortOr a = Either Abort a

Abortable type

(<!!>) :: [(String, a)] -> String -> AbortOr a

Lookup association list. This function may abort on AbortLookup.

bug :: a

Stop on error ''bug in koshucode''

Data

class IntValue v where

Methods

intValue :: Int -> v

data Rel v

Relations on type v. Heading of relation and body of relation as a list of tuples.

Constructors

Rel Relhead (Relbody v)

class RelValue v where

Methods

relValue :: Rel v -> v

data Relhead

Heading of relation as a list of terms

Constructors

Relhead

Fields

headTerms :: [Relterm]

type RelmapFun v = Rel v -> Rel v

Relation-to-relation mapping

data Relterm

Term in heading of relation

Constructors

Term String

Nest String [Relterm]

class StringValue v where

Methods

stringValue :: String -> v

headFrom :: [String] -> Relhead

Make head from term names

headPosh :: Relhead -> Relhead -> [TermPos]

Positions of given (sub)head in a head

headPoss :: Relhead -> [[String]] -> [TermPos]

Positions of given names in a head

intValue :: IntValue v => Int -> v

posPoss :: [TermPos] -> [Int]

Indicies

possInner :: [TermPos] -> [[String]]

Pick an inner part.

possOuter :: [TermPos] -> [[String]]

Pick an outer part.

possPick :: [TermPos] -> [v] -> [v]

Pick values

rehead :: ([String] -> [String]) -> Relhead -> Relhead

Reconstruct head

relValue :: RelValue v => Rel v -> v

reldee :: Rel v

Relational constant that has no terms and the empty tuple

reldum :: Rel v

Relational constant that has no terms and no tuples

stringValue :: StringValue v => String -> v

Order

orders :: String -> [OrderCap [Char]]

sortByName :: (Ord v, Eq n) => [OrderCap n] -> [n] -> [[v]] -> [[v]]

Prelude

type Map a = a -> a

type Listmap a = forall a. Map [a]

gatherToMap :: Ord k => [(k, v)] -> Map k [v]

Gather (key,value) to a Map key [value]

indexCut :: [Int] -> [a] -> [a]

Cut indexed elements

>>> indexCut [1,3] "abcdefg"
"acefg"

indexPick :: [Int] -> [a] -> [a]

Pick indexed elements.

>>> indexPick [1,3] "abcdefg"
"bd"

lookupMap :: Ord k => k -> Map k a -> Maybe a

unique :: Ord a => [a] -> [a]

Remove duplicate elements.

Relmap

type OpCons v = OpUse v -> AbortOr (Relmap v)

Constructor of relational operator Relmap. Relmap is constructed from HalfRelmap and subrelmaps in it.

data OpImplement v

Implementation of relmap operator. It consists of (1) operator name, (2) operand parser, (3) constructor of operator, and (4) usage of operator.

Constructors

OpImplement String OpParser (OpCons v) [String]

type OpParser = [TokenTree] -> [Named [TokenTree]]

Parser for operand of relational operator. This parsers docompose operand trees, and give a name to suboperand.

type OpParser' = [Named [TokenTree]] -> [Named [TokenTree]]

data OpUse v

Constructors

OpUse

Fields

opHalf :: HalfRelmap

opSub :: [Relmap v]

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a use of relational operator, that is pair of operator and its operand.

Constructors

RelmapSource HalfRelmap String [String]
Relmap that retrieves a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Relmap that has a constant relation

RelmapAlias HalfRelmap (Relmap v)
Relmap that has equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Relmap that joints two relmaps

RelmapName HalfRelmap String
Relmap reference

data RelmapCons v

Half and full relmap constructor

Constructors

RelmapCons

Fields

consHalf :: RelmapHalfCons

consFull :: RelmapFullCons v

type RelmapFullCons v

Arguments

= HalfRelmap

-> AbortOr (Relmap v)
Result relmap

Second step of constructing relmap, make Relmap from contents of HalfRelmap.

type RelmapHalfCons

Arguments

= [SourceLine]
Source information

-> [TokenTree]
Operand as source trees

-> HalfRelmap
Result half relmap

First step of constructing relmap, make HalfRelmap from use of relational operator.

halfOperand :: HalfRelmap -> [Named [TokenTree]]

Operand of relmap operation

relmapCalc :: OpUse v -> String -> RelmapSub v -> Relmap v

Make a non-confluent relmap

relmapConfl :: OpUse v -> String -> RelmapSub v -> [Relmap v] -> Relmap v

Make a confluent relmap

relmapCons

Arguments

:: [OpImplement v]
Implementations of relmap operator

-> RelmapCons v
Relmap constructors

Make half and full relmap constructors.

relmapSource :: OpUse v -> String -> [String] -> Relmap v

Retrieve relation from dataset

Section

data QuasiQuoter

koshuQuoter

Arguments

:: RelmapHalfCons
Relmap half constructor

-> ExpQ
Quotation expression of RelmapFullCons

-> QuasiQuoter
Quoter that outputs Section or Relmap

Make quasiquoter for [koshu| ... |].

Syntax

type Calc v = [v] -> v

Calculator combining values into a single value

type Ripen x v = x -> [v] -> Calc v

Set of operators

data Token

Constructors

Word Int String
Text

TermN [String]
Term name

TermP [Int]
Term position

Open String
Open paren

Close String
Close paren

Space Int
N space chars

Comment String
Comment text

Line SourceLine
Source infomation

type TokenTree = Tree Token

Tree of tokens

data Tree a

Tree of leaf and branch

Constructors

TreeL a
Terminal of a tree

TreeB Int [Tree a]
Paren-type and subtrees

binaryTree

Arguments

:: Show a

=> (a -> BinaryHeight)
Height function

-> Tree a
Infixed tree

-> Tree a
Prefixed tree

Split branches in a given tree at infixed binary operators

crop

Arguments

:: Ripen x v
Set of Operators

-> Tree x
Cropping target

-> Calc v
Calculator

Calculate a value of tree

heightTable

Arguments

:: Ord a

=> [(BinaryHeight, [a])]
Height table

-> a
Height function

-> BinaryHeight

Make the height function from a height table of operators

operandGroup :: [TokenTree] -> [Named [TokenTree]]

Split operand into named group. Non quoted words beginning with hyphen, e.g., -x, are name of group.

>>> operandGroup $ tokenTrees $ tokens "a b -x c 'd' -y e"
[("",   [TreeL (Word 0 "a"), TreeL (Word 0 "b")]),
 ("-x", [TreeL (Word 0 "c"), TreeL (Word 1 "d")]),
 ("-y", [TreeL (Word 0 "e")])]

termTreePairs :: [TokenTree] -> AbortOr [(String, TokenTree)]

Extract a list of name-and-tree pairs.

>>> termTreePairs $ tokenTrees $ tokens "/a 'A3' /b 10"
Right [("/a", TreeL (Word 1 "A3")), ("/b", TreeL (Word 0 "10"))]

WithName

withN1 :: WithTerms v

Call a relmap function with new term names.

withP :: WithTerms v

Call a relmap function with present term names.

withP2 :: ([String] -> [Rel v] -> RelmapFun v) -> String -> [Rel v] -> RelmapFun v

Call a relconfl function with present term names.