Koshucode.Baala.Minimal.Relmap

Description

Minimal relational operators.

Utility

module Koshucode.Baala.Minimal.Relmap.Get

module Koshucode.Baala.Minimal.Relmap.Pattern

module Koshucode.Baala.Minimal.Relmap.Operand

Implementation

module Koshucode.Baala.Minimal.Relmap.Tropashko

module Koshucode.Baala.Minimal.Relmap.Unary

module Koshucode.Baala.Minimal.Relmap.Restrict

module Koshucode.Baala.Minimal.Relmap.Implement

Naming conventions

Functions for relational mapping are named under some conventions. Relational mapping is a kind of functions that calculate relations from relations.

relXxx

Functions from Rel to Rel. These are basic functions for relational operators.

relmapXxx

Functions from Relmap to Relmap.

consXxx

Functions of Relop. These functions construct Relmap from operator usages, but constructions are failed if operators are misused.

likeXxx

Operand parser.

LikeXxx

Operand pattern.