Koshucode.Baala.Minimal.Relmap.Tropashko

Description

Fundamental operators in relational algebra.

Tropashko's relational lattice is a kind of relational algebra. Relational algebra is an algebraic formulation for relational model. In constrast to Codd's original relational algebra, Tropashko lattice is in more conventional and strict ways. The lattice has fundamental operators from which other operators are derived.

Fundamental operators

In lattice theory, there are two operators meet and join.

R1 ^ R2

Meet of R1 and R2. This operator is one of generalized intersection. ''Meet'' is same as ''natural join'' in SQL. R1 tuple is meetable to R2 tuple if and only if shared terms of R1 tuple and R2 tuple are same contents.

R1 v R2

Join of R1 and R2. This operator is one of generalized union. Toropashko calls it ''inner union''. Join of R1 and R2 is ordinary union of shared-term-projection tuples in R1 and R2.

Meet (Natural join)

relmapMeet

Arguments

:: Ord v

=> OpUse v
Source infomation

-> Relmap v
Subrelmap of join operator

-> Relmap v
Relmap of join operator

Meet two relations.

relMeet

Arguments

:: Ord v

=> Rel v
Input relation R1

-> Rel v
Input relation R2

-> Rel v
Meet of R1 and R2

Meet two relations.

Join (Inner union)

relmapJoin

Arguments

:: Ord v

=> OpUse v
Source infomation

-> Relmap v
Subrelmap of join operator

-> Relmap v
Relmap of join operator

Join two relations.

relJoin

Arguments

:: Ord v

=> Rel v
Input relation R1

-> Rel v
Input relation R2

-> Rel v
Join of R1 and R2

Join two relations.